Handled exceptions

+++ src/main/java/com/kingwanyama/fullstackspringbootandreact/student/StudentService.java
@@ -1,9 +1,12 @@
 package com.kingwanyama.fullstackspringbootandreact.student;
 
+import com.kingwanyama.fullstackspringbootandreact.student.exception.BadRequestException;
+import com.kingwanyama.fullstackspringbootandreact.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @AllArgsConstructor
 @Service
@@ -15,11 +18,15 @@ public class StudentService {
     }
 
     public void addStudent(Student student){
+        if (studentRepository.selectExistsEmail(student.getEmail()))
+            throw new BadRequestException("Email exists");
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long id){
-        studentRepository.deleteById(id);
+        if (studentRepository.existsById(id))
+            studentRepository.deleteById(id);
+        throw new StudentNotFoundException("Student not found");
     }
 
 }
